+++ Proyecto Spring Bruno/VF/Elchido/precaucion/RIDESCOM/src/java/mx/ipn/escom/ridescom/controller/Sedes.java
@@ -14,6 +14,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import mx.ipn.escom.ridescom.config.Conexion;
+import mx.ipn.escom.ridescom.config.Connect;
 import mx.ipn.escom.ridescom.model.Sede;
 import mx.ipn.escom.ridescom.model.Usuario;
 import mx.ipn.escom.ridescom.model.UsuarioDAO;
@@ -32,6 +33,7 @@ public class Sedes {
     UsuarioDAO udao=new UsuarioDAO();
     Usuario us=new Usuario();
     Sede pr=new Sede();
+    Connect cn=new Connect();
     int SedeID;
     
     Connection ct;
@@ -104,26 +106,52 @@ public class Sedes {
     @RequestMapping(value="DDyFD/Sedes/AgregarSede.html", method=RequestMethod.GET)
     public ModelAndView Agrega(HttpServletRequest re)throws SQLException{
         HttpSession session = re.getSession();
+        mav.clear();
         if(session.getAttribute("Nombre_U")== null){
          mav.setViewName("Error404");
         }else if(session.getAttribute("Nombre_U").equals("DDyFD")){
-        mav.setViewName("AgregarSede");
+            
+            mav.setViewName("AgregarSede");
         }else{
             mav.setViewName("CoordUA");
         }
         return mav;
     }
+    Connection se;
+    ResultSet rsed;
+    PreparedStatement stsed;
+    String exiss;
     @RequestMapping(value="DDyFD/Sedes/AgregarSede.html", method=RequestMethod.POST)
     public ModelAndView Agrega(HttpServletRequest req, HttpServletResponse resp, Sede sed)throws Exception{
+        
+    String sede=sed.getSede();
+    String sqlexis="select ID_Sede from Sede where Nombre_S='"+sede+"'";
+    try{
+            se=cn.Connect();
+            stsed=se.prepareStatement(sqlexis);
+            rsed=stsed.executeQuery();
+            if(rsed!=null ){
+                while(rsed.next() ){
+                exiss =rsed.getString("ID_Sede");
+                }
+            }
+            }catch(Exception ex){
+            }    
+        if(exiss==null){
             if(sed.getNumero().isEmpty()){
-            String sql="insert into Sede (Nombre_S, Calle, Colonia, Numero, CP, Municipio_ID_Municipio, Municipio_Estados_ID_estado) values (?, ?, ?, 's/n', ?, ?, (select Estados_ID_estado from Municipio where ID_Municipio=?))";
-            this.rid.update(sql, sed.getSede(), sed.getCalle(), sed.getColonia(), sed.getCP(), sed.getMunicipio(), sed.getMunicipio());
+                String sql="insert into Sede (Nombre_S, Calle, Colonia, Numero, CP, Municipio_ID_Municipio, Municipio_Estados_ID_estado) values (?, ?, ?, 's/n', ?, ?, (select Estados_ID_estado from Municipio where ID_Municipio=?))";
+                this.rid.update(sql, sed.getSede(), sed.getCalle(), sed.getColonia(), sed.getCP(), sed.getMunicipio(), sed.getMunicipio());
+            }else{
+                String sql="insert into Sede (Nombre_S, Calle, Colonia, Numero, CP, Municipio_ID_Municipio, Municipio_Estados_ID_estado) values (?, ?, ?, ?, ?, ?, (select Estados_ID_estado from Municipio where ID_Municipio=?))";
+                this.rid.update(sql, sed.getSede(), sed.getCalle(), sed.getColonia(), sed.getNumero(), sed.getCP(), sed.getMunicipio(), sed.getMunicipio());
+            }
+            mav.setViewName("redirect:../Sedes.html");
+            mav.addObject("msjs", "<div style='color: green;'>Se ha Agregado una sede correctamente</div>");
+
         }else{
-        String sql="insert into Sede (Nombre_S, Calle, Colonia, Numero, CP, Municipio_ID_Municipio, Municipio_Estados_ID_estado) values (?, ?, ?, ?, ?, ?, (select Estados_ID_estado from Municipio where ID_Municipio=?))";
-            this.rid.update(sql, sed.getSede(), sed.getCalle(), sed.getColonia(), sed.getNumero(), sed.getCP(), sed.getMunicipio(), sed.getMunicipio());
-        }
-        mav.setViewName("redirect:../Sedes.html");
-        mav.addObject("msjs", "<div style='color: green;'>Se ha Agregado una sede correctamente</div>");
+            mav.addObject("mjs", "<div style='color: red;'>La sede ya fue registrada, intente con otro nombre</div>");
+            mav.setViewName("redirect:../Sedes.html");
+        }    
         return mav;
     }
     @RequestMapping(value="DDyFD/Sedes/Sedesiguiente.html", method=RequestMethod.GET)
@@ -169,30 +197,66 @@ public class Sedes {
         return mav;
     }
     //Scripts para borrar registros
-    @RequestMapping(value="DDyFD/Sedes/BorrarSede.html", method=RequestMethod.GET)
-    public ModelAndView delete(HttpServletRequest re){
-        HttpSession session = re.getSession();
-        if(session.getAttribute("Nombre_U")== null){
-         mav.setViewName("Error404");
-        }else if(session.getAttribute("Nombre_U").equals("DDyFD")){
-        SedeID=Integer.parseInt(re.getParameter("SedeID"));
-        String sql="SELECT * from sede s INNER join (municipio mu, estados edo) on (mu.Estados_ID_estado=edo.ID_estado and mu.ID_Municipio=s.Municipio_ID_Municipio and s.Municipio_Estados_ID_estado=mu.Estados_ID_estado) where ID_Sede="+SedeID;
-        dat = this.rid.queryForList(sql);
-        mav.addObject("sed",dat);
-        mav.setViewName("BorrarSede");
-        }else{
-            mav.setViewName("CoordUA");
-        }
-        return mav;
-    }    
-    @RequestMapping(value="DDyFD/Sedes/BorrarSede.html", method=RequestMethod.POST)
+//    @RequestMapping(value="DDyFD/Sedes/BorrarSede.html", method=RequestMethod.GET)
+//    public ModelAndView delete(HttpServletRequest re){
+//        HttpSession session = re.getSession();
+//        if(session.getAttribute("Nombre_U")== null){
+//         mav.setViewName("Error404");
+//        }else if(session.getAttribute("Nombre_U").equals("DDyFD")){
+//        SedeID=Integer.parseInt(re.getParameter("SedeID"));
+//        String sql="SELECT * from sede s INNER join (municipio mu, estados edo) on (mu.Estados_ID_estado=edo.ID_estado and mu.ID_Municipio=s.Municipio_ID_Municipio and s.Municipio_Estados_ID_estado=mu.Estados_ID_estado) where ID_Sede="+SedeID;
+//        dat = this.rid.queryForList(sql);
+//        mav.addObject("sed",dat);
+//        mav.setViewName("BorrarSede");
+//        }else{
+//            mav.setViewName("CoordUA");
+//        }
+//        return mav;
+//    }    
+    public String sede;
+    @RequestMapping(value="DDyFD/Sedes/BorrarSede.html")
     public ModelAndView delete(HttpServletRequest re, Sede sed){
         SedeID=Integer.parseInt(re.getParameter("SedeID"));
-        String sql ="delete from Sede where ID_Sede="+SedeID;
-        this.rid.update(sql);
-//        ModelAndView mv=new ModelAndView ("redirect:../Sedes.html");
-        mav.setViewName("redirect:../Sedes.html");
-        mav.addObject("msjs", "<div style='color: green;'>Se ha eliminado correctamente</div>");
+        String sql="select distinct i.Evento_Evento_ID from Inscripcion i inner join (act_deportiva dep, pruebas pr, evento ev, Sede se) on (dep.ID_Deporte=pr.Act_Deportiva_ID_Deporte and pr.ID_Pruebas=ev.Pruebas_ID_Pruebas and ev.Evento_ID=i.Evento_Evento_ID and se.ID_Sede=ev.Sede_ID_Sede) where se.ID_Sede="+SedeID;
+        java.sql.Connection ctd;
+        ResultSet rsd;
+        PreparedStatement psd;
+        String de=null;
+        try{
+            ctd=cn.Connect();
+            psd=ctd.prepareStatement(sql);
+            rsd=psd.executeQuery();
+            if(rsd!=null ){
+                while(rsd.next() ){
+                de =rsd.getString("Evento_Evento_ID");
+                }
+            }
+        }catch(Exception e){
+        }
+        if(de==null){
+            String sqlse ="delete from Sede where ID_Sede="+SedeID;
+            this.rid.update(sqlse);
+            mav.addObject("msjs", "<div style='color: green;'>Se ha eliminado correctamente</div>");
+            mav.setViewName("redirect:../Sedes.html");
+        }else{
+            SedeID=Integer.parseInt(re.getParameter("SedeID"));
+            String sqlds="SELECT Nombre_S from sede where ID_Sede="+SedeID;
+            try{
+            se=cn.Connect();
+            stsed=se.prepareStatement(sqlds);
+            rsed=stsed.executeQuery();
+            if(rsed!=null ){
+                while(rsed.next() ){
+                sede =rsed.getString("Nombre_S");
+                }
+            }
+        }catch(Exception e){
+        }
+            dat = this.rid.queryForList(sqlds);
+            mav.addObject("sed",dat);
+            mav.addObject("msjs","<div style='color: red;'>La sede \""+sede+"\" ya alberga un evento con alumnos inscritos</div>");
+            mav.setViewName("redirect:../Sedes.html");
+        }
         return mav;
     }
     //No sirve más
